# Conflicts:
#	buildSrc/build.gradle.kts
#	buildSrc/settings.gradle.kts
#	gradle.properties
#	gradle/libs.versions.toml
#	kotlin-js-store/yarn.lock

+++ ktor-server/ktor-server-config-yaml/nix/test/YamlConfigTestNix.kt
@@ -82,7 +82,7 @@ class YamlConfigTestNix {
         val file = fopen(path, "w")
         val bytes = content.toByteArray()
         bytes.usePinned { pinned ->
-            fwrite(pinned.addressOf(0), 1, bytes.size.convert(), file)
+            fwrite(pinned.addressOf(0), 1.toULong(), bytes.size.convert(), file)
         }
         fclose(file)
     }
